+++ src/main/java/org/opensearch/knn/index/codec/nativeindex/remote/RemoteIndexBuildStrategy.java
@@ -11,10 +11,13 @@ import org.apache.lucene.index.SegmentWriteState;
 import org.opensearch.common.StopWatch;
 import org.opensearch.common.annotation.ExperimentalApi;
 import org.opensearch.index.IndexSettings;
+import org.opensearch.knn.common.KNNConstants;
 import org.opensearch.knn.common.featureflags.KNNFeatureFlags;
 import org.opensearch.knn.index.KNNSettings;
 import org.opensearch.knn.index.codec.nativeindex.NativeIndexBuildStrategy;
 import org.opensearch.knn.index.codec.nativeindex.model.BuildIndexParams;
+import org.opensearch.knn.index.remote.RemoteBuildRequest;
+import org.opensearch.knn.index.remote.RemoteIndexClient;
 import org.opensearch.knn.index.vectorvalues.KNNVectorValues;
 import org.opensearch.repositories.RepositoriesService;
 import org.opensearch.repositories.Repository;
@@ -22,10 +25,13 @@ import org.opensearch.repositories.RepositoryMissingException;
 import org.opensearch.repositories.blobstore.BlobStoreRepository;
 
 import java.io.IOException;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.function.Supplier;
 
 import static org.opensearch.knn.index.KNNSettings.KNN_INDEX_REMOTE_VECTOR_BUILD_SETTING;
 import static org.opensearch.knn.index.KNNSettings.KNN_REMOTE_VECTOR_REPO_SETTING;
+import static org.opensearch.knn.index.KNNSettings.state;
 
 /**
  * This class orchestrates building vector indices. It handles uploading data to a repository, submitting a remote
@@ -54,7 +60,7 @@ public class RemoteIndexBuildStrategy implements NativeIndexBuildStrategy {
      * @return whether to use the remote build feature
      */
     public static boolean shouldBuildIndexRemotely(IndexSettings indexSettings) {
-        String vectorRepo = KNNSettings.state().getSettingValue(KNN_REMOTE_VECTOR_REPO_SETTING.getKey());
+        String vectorRepo = state().getSettingValue(KNN_REMOTE_VECTOR_REPO_SETTING.getKey());
         return KNNFeatureFlags.isKNNRemoteVectorBuildEnabled()
             && indexSettings.getValue(KNN_INDEX_REMOTE_VECTOR_BUILD_SETTING)
             && vectorRepo != null
@@ -88,17 +94,18 @@ public class RemoteIndexBuildStrategy implements NativeIndexBuildStrategy {
             log.debug("Repository write took {} ms for vector field [{}]", time_in_millis, indexInfo.getFieldName());
 
             stopWatch = new StopWatch().start();
-            submitVectorBuild();
+            RemoteBuildRequest buildRequest = constructBuildRequest(indexInfo);
+            String jobId = RemoteIndexClient.getInstance().submitVectorBuild(buildRequest);
             time_in_millis = stopWatch.stop().totalTime().millis();
             log.debug("Submit vector build took {} ms for vector field [{}]", time_in_millis, indexInfo.getFieldName());
 
             stopWatch = new StopWatch().start();
-            awaitVectorBuild();
+            String indexPath = awaitVectorBuild(jobId);
             time_in_millis = stopWatch.stop().totalTime().millis();
             log.debug("Await vector build took {} ms for vector field [{}]", time_in_millis, indexInfo.getFieldName());
 
             stopWatch = new StopWatch().start();
-            readFromRepository();
+            readFromRepository(indexPath);
             time_in_millis = stopWatch.stop().totalTime().millis();
             log.debug("Repository read took {} ms for vector field [{}]", time_in_millis, indexInfo.getFieldName());
         } catch (Exception e) {
@@ -146,6 +153,49 @@ public class RemoteIndexBuildStrategy implements NativeIndexBuildStrategy {
     }
 
     /**
+     * Construct the RemoteBuildRequest object for the index build request
+     * @return RemoteBuildRequest with parameters set
+     */
+    public RemoteBuildRequest constructBuildRequest(BuildIndexParams indexInfo) throws IOException {
+        String repositoryType = getRepository().getMetadata().type();
+        String containerName = switch (repositoryType) {
+            case "s3" -> getRepository().getMetadata().settings().get("bucket");
+            case "fs" -> getRepository().getMetadata().settings().get("location");
+            default -> throw new IllegalStateException("Unexpected value: " + repositoryType);
+        };
+        String vectorPath = null; // blobName + VECTOR_BLOB_FILE_EXTENSION
+        String docIdPath = null; // blobName + DOC_ID_FILE_EXTENSION
+        String tenantId = null; // indexSettings.getSettings().get(ClusterName.CLUSTER_NAME_SETTING.getKey());
+        int dimension = 0; // TODO
+        int docCount = indexInfo.getTotalLiveDocs();
+        String dataType = indexInfo.getVectorDataType().getValue(); // TODO need to fetch encoder param to get fp16 vs fp32
+        String engine = indexInfo.getKnnEngine().getName();
+
+        String spaceType = indexInfo.getParameters().get(KNNConstants.SPACE_TYPE).toString();
+
+        Map<String, Object> algorithmParams = new HashMap<>();
+        algorithmParams.put("ef_construction", 100);
+        algorithmParams.put("m", 16);
+
+        Map<String, Object> indexParameters = new HashMap<>();
+        indexParameters.put("algorithm", "hnsw");
+        indexParameters.put("algorithm_parameters", algorithmParams);
+
+        return RemoteBuildRequest.builder()
+            .repositoryType(repositoryType)
+            .containerName(containerName)
+            .vectorPath(vectorPath)
+            .docIdPath(docIdPath)
+            .tenantId(tenantId)
+            .dimension(dimension)
+            .docCount(docCount)
+            .dataType(dataType)
+            .engine(engine)
+            .indexParameters(indexParameters)
+            .build();
+    }
+
+    /**
      * Submit vector build request to remote vector build service
      *
      */
@@ -156,14 +206,14 @@ public class RemoteIndexBuildStrategy implements NativeIndexBuildStrategy {
     /**
      * Wait on remote vector build to complete
      */
-    private void awaitVectorBuild() {
+    private String awaitVectorBuild(String jobId) {
         throw new NotImplementedException();
     }
 
     /**
      * Read constructed vector file from remote repository and write to IndexOutput
      */
-    private void readFromRepository() {
+    private void readFromRepository(String indexPath) {
         throw new NotImplementedException();
     }
 }